remove tasks 1 to 3
