+++ src/1/index.scss
@@ -1,4 +1,4 @@
 /* DO NOT CHANGE THIS ROOT CLASS NAME*/
-#task-6 {
+#task-1 {
   /* Styles goes here */
 }
