+++ src/3/index.scss
@@ -1,4 +1,4 @@
 /* DO NOT CHANGE THIS ROOT CLASS NAME*/
-#task-5 {
+#task-3 {
   /* Styles goes here */
 }
