--- src/5/index.tsx
@@ -1,17 +1,0 @@
-import { FunctionComponent } from "react";
-
-// Components
-import Input from "./components/Input";
-import List from "./components/List";
-
-const Task4: FunctionComponent = () => {
-  return (
-    <div>
-      <Input />
-      <br />
-      <List />
-    </div>
-  );
-};
-
-export default Task4;
