--- src/6/index.tsx
@@ -1,16 +1,0 @@
-import * as React from "react";
-
-// Style
-import "./index.scss";
-
-// Components
-/*
- * Create the components you need in the components folder.
- * You may find inspiration in task 4
- */
-
-const Task6: React.FunctionComponent = () => {
-  return <div id="task-6">#Code goes here#</div>;
-};
-
-export default Task6;
