+++ src/index.tsx
@@ -14,12 +14,10 @@ import {
 } from "react-router-dom";
 import Intro from "./Intro";
 import { NavLink } from "react-router-dom";
+
 import Task1 from "./1";
 import Task2 from "./2";
 import Task3 from "./3";
-import Task4 from "./4";
-import Task5 from "./5";
-import Task6 from "./6";
 
 const Menu: FunctionComponent = () => {
   return (
@@ -31,12 +29,6 @@ const Menu: FunctionComponent = () => {
       <NavLink to="/task/2">Task 2</NavLink>
 
       <NavLink to="/task/3">Task 3</NavLink>
-
-      <NavLink to="/task/4">Task 4</NavLink>
-
-      <NavLink to="/task/5">Task 5</NavLink>
-
-      <NavLink to="/task/6">Task 6</NavLink>
     </nav>
   );
 };
@@ -66,58 +58,10 @@ const Task: FunctionComponent = () => {
   let { id } = useParams<{ id: string }>();
 
   switch (id) {
-    case "1":
-      return (
-        <>
-          <h1>Task 1 - Create a list</h1>
-          <p className="task-description">
-            This first task is purely about using JSX. Make an ordinary
-            unordered list using 'ul' and li'. The list must{" "}
-            <b>not be static</b>. You have to <b>use the array</b> in state and{" "}
-            <b>iterate over</b> it.
-          </p>
-
-          <TaskContainer>
-            <Task1 />
-          </TaskContainer>
-        </>
-      );
-    case "2":
-      return (
-        <>
-          <h1>Task 2 - Create a controlled input</h1>
-          <p className="task-description">
-            This task is purely about using <b>React's synthetic event</b>{" "}
-            system. Use the existing input and paragraph to display the{" "}
-            <b>the input value</b> in the paragraph. The paragraph must be kept
-            in sync.
-          </p>
-
-          <TaskContainer>
-            <Task2 />
-          </TaskContainer>
-        </>
-      );
-    case "3":
-      return (
-        <>
-          <h1>Task 3 - Create a searchable list</h1>
-          <p className="task-description">
-            Use the <b>knowledge</b> you used in the <b>previous 2 tasks</b> to
-            make at searchable list. You must have a <b>input</b> and a{" "}
-            <b>list</b>. The list should be <b>filtered</b> using the value in
-            the input as the user types.
-          </p>
-
-          <TaskContainer>
-            <Task3 />
-          </TaskContainer>
-        </>
-      );
-      case "4":
+      case "1":
         return (
           <>
-            <h1>Task 4 - Use styles</h1>
+            <h1>Task 1 - Use styles</h1>
             <p className="task-description">
               This task is mainly about <b>styling</b> in ReactJS in order to
               create a <b>beautiful UI</b>. Use Less to style the form with
@@ -126,14 +70,14 @@ const Task: FunctionComponent = () => {
             </p>
   
             <TaskContainer>
-              <Task4 />
+              <Task1 />
             </TaskContainer>
           </>
         );
-      case "5":
+      case "2":
       return (
         <>
-          <h1>Task 5 - Use component composition</h1>
+          <h1>Task 2 - Use component composition</h1>
           <p className="task-description">
             Use the <b>knowledge</b> you used in the <b>previous tasks</b> to
             make a searchable list. The requirement is the same as with Task 3.
@@ -147,14 +91,14 @@ const Task: FunctionComponent = () => {
           </p>
 
           <TaskContainer>
-            <Task5 />
+            <Task2 />
           </TaskContainer>
         </>
       );
-    case "6":
+    case "3":
       return (
         <>
-          <h1>Task 6 - Create a Todo List and/or Freestyle</h1>
+          <h1>Task 3 - Create a Todo List and/or Freestyle</h1>
           <p className="task-description">
             This task is mainly about <b>combining different concepts</b> from
             ReactJS in order to create a <b>small and simple application</b>.
@@ -173,7 +117,7 @@ const Task: FunctionComponent = () => {
           </p>
 
           <TaskContainer>
-            <Task6 />
+            <Task3 />
           </TaskContainer>
         </>
       );
